Master

@wlee @achavan changes pushed to qa maybe make the 1200 qa push

See merge request !4

+++ persistence-jpa/src/main/java/com/emmisolutions/emmimanager/persistence/repo/ProgramRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 /**
  * A Program spring data repository
  */
-public interface ProgramRepository extends JpaRepository<Program, Long>, JpaSpecificationExecutor<Program> {
+public interface ProgramRepository extends JpaRepository<Program, Integer>, JpaSpecificationExecutor<Program> {
 }
